feat(DebugInfo): add DebugInfo component
- create a core.gfx package,
- add a new DebugInfo class,
- move the Renderer class to the gfx package

Fix #30
