--- src/main/resources/res/assets/asset-1.json
@@ -1,192 +1,0 @@
-{
-  "name": "Objects Set 1",
-  "image": "/res/images/tileset-1.png",
-  "tileWidth": 16,
-  "tileHeight": 16,
-  "objects": {
-    "E": {
-      "id": "E",
-      "type": "enemy_",
-      "clazz": "core.object.GameObject",
-      "image": "todo",
-      "offset": "6,1",
-      "size": "32,32",
-      "priority": 2,
-      "layer": 1,
-      "bbox": {
-        "top": 8.0,
-        "bottom": 0.0,
-        "left": 4.0,
-        "right": 4.0
-      },
-      "color": "RED",
-      "canCollect": false,
-      "attributes": {
-        "energy": 100.0,
-        "damage": 10.0,
-        "coins": 5
-      }
-    },
-    "P": {
-      "id": "P",
-      "type": "player",
-      "clazz": "core.object.GameObject",
-      "image": "todo",
-      "offset": "4,1",
-      "size": "32,32",
-      "priority": 10,
-      "layer": 1,
-      "bbox": {
-        "top": 8.0,
-        "bottom": 0.0,
-        "left": 4.0,
-        "right": 4.0
-      },
-      "color": "BLUE",
-      "canCollect": true,
-      "attributes": {
-        "energy": 100.0,
-        "mana": 100.0,
-        "damage": 50.0,
-        "coins": 250,
-        "items": [
-          "sword-15",
-          "nrj+10"
-        ],
-        "maxItems": 3
-      }
-    },
-    "1": {
-      "type": "item",
-      "id": "1",
-      "description": "sword-1",
-      "usage": "weapon",
-      "collectible": true,
-      "damage": 10.0,
-      "offset": "7,1"
-    },
-    "2": {
-      "type": "item",
-      "id": "2",
-      "description": "nrj-10",
-      "usage": "health",
-      "collectible": true,
-      "energy": 10.0,
-      "offset": "4,1"
-    },
-    "3": {
-      "type": "item",
-      "id": "3",
-      "description": "arc-10",
-      "usage": "weapon",
-      "collectible": true,
-      "energy": 10.0,
-      "offset": "5,1"
-    },
-    "o": {
-      "type": "object",
-      "id": "o",
-      "collectible": true,
-      "money": 10,
-      "offset": "10,2"
-    },
-    "O": {
-      "type": "object",
-      "id": "O",
-      "collectible": true,
-      "money": 100,
-      "offset": "11,2"
-    },
-    ">": {
-      "type": "tile",
-      "id": ">",
-      "image": "pic-right.png",
-      "hit": true,
-      "damage": 10
-    },
-    "<": {
-      "type": "tile",
-      "id": "<",
-      "image": "pic-left.png",
-      "hit": true,
-      "damage": 10
-    },
-    "V": {
-      "type": "tile",
-      "id": "V",
-      "image": "pic-down.png",
-      "hit": true,
-      "damage": 10
-    },
-    "|": {
-      "type": "tile",
-      "id": "|",
-      "image": "pic-up.png",
-      "hit": true,
-      "damage": 10
-    },
-    "b": {
-      "type": "tile",
-      "id": "b",
-      "block": true,
-      "offset": "17,8"
-    },
-    "B": {
-      "type": "tile",
-      "id": "B",
-      "block": true,
-      "offset": "19,8"
-    },
-    "H": {
-      "type": "tile",
-      "id": "H",
-      "block": false,
-      "offset": "13,6",
-      "climbable": true
-    },
-    "h": {
-      "type": "tile",
-      "id": "h",
-      "block": false,
-      "offset": "19,5",
-      "climbable": true
-    },
-    "T": {
-      "type": "tile",
-      "id": "T",
-      "block": false,
-      "offset": "19,6"
-    },
-    "k": {
-      "type": "tile",
-      "id": "k",
-      "block": false,
-      "offset": "19,5"
-    },
-    "=": {
-      "type": "tile",
-      "id": "=",
-      "block": true,
-      "offset": "19,5"
-    },
-    "X": {
-      "type": "tile",
-      "id": "X",
-      "image": "door-level-out.png",
-      "levelOutput": true,
-      "nextLevel": "map_2"
-    },
-    "(": {
-      "type": "tile",
-      "id": "(",
-      "block": true,
-      "offset": "19,7"
-    },
-    ")": {
-      "type": "tile",
-      "id": "(",
-      "block": true,
-      "offset": "20,7"
-    }
-  }
-}
