--- src/main/resources/res/maps/map_1.json
@@ -1,50 +1,0 @@
-{
-  "name": "Level 1.1",
-  "description": "This a demonstration level.",
-  "objects": "/res/assets/asset-1.json",
-  "background": "/res/images/background-1.jpg",
-  "map": {
-    "front": [
-      "BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB",
-      ")()()()()()BBbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbBBBBBBBBBBBBB",
-      "()()()Bbbbbb                   V                                   BBBBBBBBBBBBB",
-      ")()()()Bbbb                                                            BBBBBBBBB",
-      "()()()BBbbb                                                             BBBBBBBB",
-      "BBBBBBBBbbb                                                                BBBBB",
-      "BBBBBBBBbbb                                                                  BBB",
-      "BBBBBBBBbbb                                                                   BB",
-      "BBBBBBBBbbb  P     o           o      1                                       BB",
-      "BBBBBBBBBBBTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTT     [        o   E    o   ]BB",
-      "BBBbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb     TTTTTTTTTTTTTTTTTTTTTTTBB",
-      "BBb>                                                  BBBBBBBBBBBBBBBBBBBBBBBBBB",
-      "BBb>  oooo         o     o                [    E     ]BBBBBBBBBBBBBBBBBBBBBBBBBB",
-      "BBBTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTT        BhTTTTTTTTTTBBBBBBBBBBBBBBBBBBBBBBBBBB",
-      "BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBb    TTTTBHBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB",
-      "BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBb    BBBBBHBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB",
-      "BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBb    BbbbbHbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbBBBBB",
-      "BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBb         H                              <BBBBB",
-      "BBbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb         H                              <BBBBB",
-      "BBb         2             VVVV             H              TThTTTTTTTTTTTTTTBBBBB",
-      "BBb    TTTTTTTTTTTT                        H                H         BBBBBBBBBB",
-      "BBb    BBBBBBBBBBBb                        H     o      o   H   o     BBBBBBBBBB",
-      "BBb    BBBBBBBBBBBBBTTTTTThTTTTTTT    TTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTBBBBBBBBBB",
-      "BBb    VBBBBBBBBBBBBBBBBBBHBBBBBBb    BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB",
-      "BBb                       H  BBBBb    BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB",
-      "BBb                       H  BBBBb    BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB",
-      "BBBTTTTTTTT    =====k=====k==BBBBb    BbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbBBBBBBBBBB",
-      "BBBBBBBBBBb         H        BBBBb                                    VBBBBBBBBB",
-      "BBbbbbbbbbb         H        BBBBb |                                   BBBBBBBBB",
-      "Bb            ==k============BBBBBTTTTTTTTTTTTTTTTTT | TTTTTTTTTTTT    BBBBBBBBB",
-      "Bb  [E        ] H            BbbBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBb    BBBBBBBBB",
-      "Bb==============k========       BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBb    BBBBBBBBB",
-      "Bb              H               BBBBBBBBBBBBBb   BBBBBBBBBBBBBBBBBb    BBBBBBBBB",
-      "Bb [oOoOoO      H 3        E]   BBBBBBBBBBBBBb    BBBBBBBBBBBBBBBBb    BBBBBBBBB",
-      "BBTTTTTTTTTTTTTTTTTTTTTTTTTTTTTBBBBBBBBBBBb        BBBBBBBBBBBBBBBb    BBBBBBBBB",
-      "BBBBBBBBBBbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb            Bbbbbbbbbbbbb    BbbbbbbBB",
-      "BBBBBBBBBBb>                                                                 <BB",
-      "BBBBBBBBBBb X[          O           E       ]                        |       <BB",
-      "BBBBBBBBBBBTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTT ||| TTTTTT ||| TTTTTTTTTTTTTTTTTBB",
-      "BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB"
-    ]
-  }
-}
